Move search space validation tests to tests/validation/
